move extractFilesFromAzure to BaseDriver from Snowflake
