+++ packages/cubejs-base-driver/package.json
@@ -29,6 +29,7 @@
     "dist/src/*"
   ],
   "dependencies": {
+    "@azure/storage-blob": "^12.9.0",
     "@aws-sdk/client-s3": "^3.49.0",
     "@aws-sdk/s3-request-presigner": "^3.49.0",
     "@cubejs-backend/shared": "^0.36.0",
